Refactored fetching html content from the static files. The logic was moved into the common service.
